Formated the design so that the girl with the ballon takes up the full width, and adjusted ballon page position accrodingly, along with making ballon location reponsive with %of VW as apose to px.
Changed background image to a white brick wall to relfect more of Banksy's style.

Moved the information boxes to the bottom of the page, and ajusted CSS accordingly. (moved HTML div class="infobox" below ticke.)* only change to html.

Styling was all changed via a seperate CSS file, using George's CSS main as the baseline, and adjusting accrodingly. You can switch between style for the site, by just swithing which CSS style file is linked to the html.

I also adjusted one small thing on the JS file so that the ticker updates the information based on the new city clicked. The ticker was previously only updating for the first click, and then staying the same.